requirements.txt
# Enhanced Microcap Trading System Dependencies
# Production-ready requirements

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Financial data APIs
yfinance>=0.2.0
requests>=2.28.0

# Rich terminal interface
rich>=12.0.0

# Environment variable management
python-dotenv>=0.19.0

# Optional: For enhanced data sources
# These are optional - system works without them
# polygon-api-client>=1.0.0  # Uncomment if using Polygon.io
# finnhub-python>=2.4.0      # Uncomment if using Finnhub 